Create README.md
In an era where fire safety remains a paramount concern across various environments, harnessing the potential of advanced technologies is imperative for proactive threat mitigation. This project embarks on a critical mission, employing the YOLOv8 object detection algorithm, renowned for its efficiency and accuracy, to tackle the formidable challenge of fire and smoke detection. Through meticulous data curation and model training, we have endeavored to develop a real-time system capable of 
swiftly and accurately identifying these crucial indicators of potential danger. The integration of YOLOv8, a state-of-the-art object detection framework, promises to usher in a new era of fire and smoke detection capabilities, elevating safety measures 
across industrial, residential, and public domains. This report chronicles our journey, detailing the methodology, implementation, and promising results that underscore 
the potential impact of this technology on fire safety practices.